pyproject.toml
[project]
name = "project-artemis"
version = "0.1.0"
description = "Multi-agent orchestration system for job hunting and career development"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Alex Jansen" }]

dependencies = [
    # LLM & Agent Framework
    "langgraph>=0.3.0",
    "langchain>=0.3.0",
    "langchain-google-genai>=2.0.0",
    "google-generativeai>=0.8.0",
    # Vector Database
    "chromadb>=0.5.0",
    # Database
    "supabase>=2.0.0",
    # Web Framework (API layer for frontend + bot)
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Data Ingestion
    "httpx>=0.28.0",
    "feedparser>=6.0.0",
    "beautifulsoup4>=4.12.0",
    "python-docx>=1.1.0",
    # Utilities
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.0.0",
    "duckduckgo-search>=8.1.1",
    "googlesearch-python>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
discord = [
    "discord-py>=2.4.0",
]
pdf = [
    "weasyprint>=62.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "ingestion", "bot"]
